+++ scripts/erp.py
@@ -10,7 +10,7 @@ event markers of specified labels with padding intervals `left-padding` and
 argument is provided, bad segments are dropped based on the specified
 peak-to-peak amplitude criterion.  Averaging is performed per label.
 Then, the data are re-referenced to an average reference if `average-ref` flag
-is present.  These are then written to the output file path as an .mff.
+is present.  The averages are then written to the output file path as an .mff.
 """
 
 from collections import defaultdict
@@ -20,14 +20,13 @@ from functools import partial
 from typing import List, Union
 from xml.etree.ElementTree import parse
 
-from mffpy import Reader, XML
+from mffpy import XML
 from mffpy.xml_files import EventTrack
 import pytz
 
-from eegwlib.filter import filtfilt
-from eegwlib.segment import slice_block
 from eegwlib.artifact_detection import detect_bad_channels
 from eegwlib.average import Averages
+from eegwlib.segment import seconds_to_samples, Segmenter
 
 from argparse import ArgumentParser
 
@@ -136,7 +135,7 @@ if len(opt.labels) != len(category_names):
 categories = dict(zip(category_names, opt.labels))
 
 # Read raw input file
-raw = Reader(opt.input_file)
+raw = Segmenter(opt.input_file)
 sampling_rate = raw.sampling_rates['EEG']
 
 # Get history info if present
@@ -145,40 +144,6 @@ if 'history.xml' in raw.directory.listdir():
         history = XML.from_file(fp).entries
 else:
     history = []
-
-# Get raw signal blocks and apply filter if specified
-filter_start = pytz.utc.localize(datetime.utcnow())
-data = []
-for epoch in raw.epochs:
-    signals = raw.get_physical_samples_from_epoch(epoch)['EEG'][0]
-    filtered_signals = filtfilt(signals, order=opt.filter_order,
-                                sr=sampling_rate, fmin=opt.highpass,
-                                fmax=opt.lowpass)
-    data.append(
-        {
-            't0': epoch.t0,
-            't1': epoch.t1,
-            'data': filtered_signals
-        }
-    )
-filter_end = pytz.utc.localize(datetime.utcnow())
-
-for filt, freq in {'Highpass': opt.highpass, 'Lowpass': opt.lowpass}.items():
-    if freq:
-        filter_entry = dict(
-            name=f'ERP Workflow {filt} Filter',
-            kind='Transformation',
-            method='Filtering',
-            beginTime=filter_start.astimezone(opt.timezone),
-            endTime=filter_end.astimezone(opt.timezone),
-            sourceFiles=[opt.input_file],
-            settings=[f'Filter Setting: {freq} Hz {filt}',
-                      'Filter Type: IIR Butterworth',
-                      f'Filter Order: {opt.filter_order}']
-        )
-        if opt.verbose:
-            print(filter_entry)
-        history.append(filter_entry)
 
 # Extract relative times for events of specified labels
 all_codes = set()
@@ -203,32 +168,12 @@ for cat, label in categories.items():
                          f'Valid event labels: {all_codes}')
     times_by_category[cat] = sorted(event_times[label])
 
-# Extract data segments
+# Extract data segments from filtered epochs
 segment_start = pytz.utc.localize(datetime.utcnow())
-out_of_bounds_segs = defaultdict(list)
-segments = defaultdict(list)
-for cat, times in times_by_category.items():
-    block_idx = 0
-    block = data[block_idx]
-    for time in times:
-        while time > block['t1']:
-            # Iterate through data blocks until we
-            # find the one that contains time
-            block_idx += 1
-            block = data[block_idx]
-        assert block['t0'] < time < block['t1']
-        # Extract the segment centered on time
-        segment = slice_block(
-            block['data'],
-            center=time - block['t0'],
-            padl=opt.left_padding,
-            padr=opt.right_padding,
-            sr=sampling_rate
-        )
-        if segment is not None:
-            segments[cat].append(segment)
-        else:
-            out_of_bounds_segs[cat].append(time)
+segments, out_of_bounds_segs = raw.extract_segments(
+    times_by_category, opt.left_padding, opt.right_padding,
+    order=opt.filter_order, fmin=opt.highpass, fmax=opt.lowpass
+)
 segment_end = pytz.utc.localize(datetime.utcnow())
 
 # Check if any categories have no segments
@@ -237,6 +182,25 @@ for cat in categories:
         raise ValueError(f'All segments for category "{cat}" '
                          'extended beyond data range')
 
+# Log filtering
+for filt, freq in {'Highpass': opt.highpass, 'Lowpass': opt.lowpass}.items():
+    if freq:
+        filter_entry = dict(
+            name=f'ERP Workflow {filt} Filter',
+            kind='Transformation',
+            method='Filtering',
+            beginTime=segment_start.astimezone(opt.timezone),
+            endTime=segment_end.astimezone(opt.timezone),
+            sourceFiles=[opt.input_file],
+            settings=[f'Filter Setting: {freq} Hz {filt}',
+                      'Filter Type: IIR Butterworth',
+                      f'Filter Order: {opt.filter_order}']
+        )
+        if opt.verbose:
+            print(filter_entry)
+        history.append(filter_entry)
+
+# Log segmentation
 segmentation_settings = []
 segmentation_results = [f'Segmented to {len(segments)} categories and '
                         f'{sum(map(len, segments.values()))} segments']
@@ -327,8 +291,9 @@ else:
 
 # Create averages for each label
 average_start = pytz.utc.localize(datetime.utcnow())
-averages = Averages(center=int(opt.left_padding * sampling_rate),
-                    sr=sampling_rate, bads=bad_channels)
+# Snap to nearest sample
+center = seconds_to_samples(opt.left_padding, sampling_rate) / sampling_rate
+averages = Averages(center, sampling_rate, bads=bad_channels)
 for cat, data in segments.items():
     averages.add(cat, data)
 average_end = pytz.utc.localize(datetime.utcnow())