+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+* Class `Segmenter` which adds segmentation capability to parent class `mffpy.reader`
+
+### Changed
+* Fixed bug in `eegwlib.filter.butter` to provide missing sampling rate arg
 
 ## [0.1.2] - 2021-04-19
 ### Added